Add ability to edit tasks prior to save.
The previous per-task callback mechanism (executions.cb) did not
actually allow the mutation of a task's members that changed in the
synchronization. This change renames the previous callback mechanism to
'executions.filter', which continues to allow the mutation/filtering of
a task prior to sync.

This change also repurposes the old callback, as 'executions.cb', to
allow for pre-save mutations. Filtering is no longer applied at this
time.